Add files via upload
Use Haarcascade file of Face and Eye detection . Here we detect the the face and eye in the image of single person and multiple person. Also detect the face and eye of a person in real time using webcam. However this technique is old and not so smart and intelligent so  its detection accuracy is not always correct . Sometimes it gives extra detection of the eyes and face in  the object .